Synchronous Driver Actions Solution to README
Feel many people will require synchronous solution to requests prior to response parsing to ensure best compatibility with scrapy design which keeps response parsing and request asynchronous. This utilizes the wait_until to perform any driver actions prior to response formation. Could be useful to people so maybe include as use case in README? Also, there may be potential to create a wrapper that does this more user friendly, but for now this good maybe.